+++ packages/selenium/src/Resources/storeChunk.js
@@ -1,0 +1,2 @@
+window.partialResults ??= '';
+window.partialResults += arguments[0];